requirements.txt
# Dependencias Principales
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Utilidades
python-decouple==3.8
pytz==2024.1

# Cliente HTTP y Base de Datos (Opcional)
requests==2.31.0
psycopg2-binary==2.9.9

# Estándar de Respuesta
six==1.16.0